refactor(file): Move FileHash to its own file.
